refactor common methods into ContextBuilder, wow that was a blunder
