+++ src/main/java/net/exploitables/slashlib/context/ContextBuilder.java
@@ -1,8 +1,10 @@
 package net.exploitables.slashlib.context;
 
+import discord4j.core.event.domain.interaction.InteractionCreateEvent;
 import discord4j.core.object.entity.Guild;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.User;
+import discord4j.core.object.entity.channel.GuildChannel;
 import discord4j.core.object.entity.channel.MessageChannel;
 import discord4j.core.object.entity.channel.TopLevelGuildChannel;
 import reactor.core.publisher.Mono;
@@ -26,6 +28,8 @@ import java.util.List;
  * It is a bit messy, but results in consistent access during the command lifecycle.
  */
 public abstract class ContextBuilder {
+    protected final @NonNull InteractionCreateEvent event;
+    
     // List of Monos which will be zipped when building to gather all required data, throws an exception if empty
     protected final List<Mono<Integer>> requiredMonoList;
     // List of Monos which will be zipped when building to gather all optional data, doesn't throw an exception if empty
@@ -45,9 +49,11 @@ public abstract class ContextBuilder {
      * Create a new context builder, the user must be specified as all
      *  interaction events provide the user through D4J.
      *
-     * @param user the user who called the interaction
+     * @param event the event related to this context
      */
-    ContextBuilder(@NonNull User user) {
+    ContextBuilder(@NonNull InteractionCreateEvent event) {
+        this.event = event;
+        
         this.requiredMonoList = new ArrayList<>();
         this.requestMonoList = new ArrayList<>();
         this.allRequestedDataExists = false;
@@ -63,7 +69,7 @@ public abstract class ContextBuilder {
         this.guild = null;
         this.messageChannel = null;
         this.topLevelGuildChannel = null;
-        this.user = user;
+        this.user = event.getInteraction().getUser();
         this.member = null;
         this.botUser = null;
         this.botMember = null;
@@ -91,18 +97,195 @@ public abstract class ContextBuilder {
             .thenReturn(this);
     }
 
-    // Provide these methods as abstract as they must be implemented by
-    //  subclasses since the methods to get the data isn't common between events
-    public abstract ContextBuilder requireGuild();
-    public abstract ContextBuilder requestGuild();
-    public abstract ContextBuilder requireMessageChannel();
-    public abstract ContextBuilder requestMessageChannel();
-    public abstract ContextBuilder requireTopLevelGuildChannel();
-    public abstract ContextBuilder requestTopLevelGuildChannel();
-    public abstract ContextBuilder requireMember();
-    public abstract ContextBuilder requestMember();
-    public abstract ContextBuilder requireBotUser();
-    public abstract ContextBuilder requestBotUser();
-    public abstract ContextBuilder requireBotMember();
-    public abstract ContextBuilder requestBotMember();
+    /**
+     * Mark that the {@link Guild} the command was called in is required for command execution.
+     * This method will collect the:
+     * guild
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requireGuild() {
+        requiredMonoList.add(event.getInteraction().getGuild()
+            .doOnNext(guild -> this.guild = guild)
+            .map(guild -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link Guild} the command was called in is optionally used for command execution.
+     * This method will collect the:
+     * guild
+     *
+     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
+     * @return this instance
+     */
+    public ContextBuilder requestGuild() {
+        requestMonoList.add(event.getInteraction().getGuild()
+            .doOnNext(guild -> this.guild = guild)
+            .map(guild -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link MessageChannel} the command was called in is required for command execution.
+     * This method will collect the:
+     * messageChannel
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requireMessageChannel() {
+        requiredMonoList.add(event.getInteraction().getChannel()
+            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
+            .map(messageChannel -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link MessageChannel} the command was called in is optionally used for command execution.
+     * This method will collect the:
+     * messageChannel
+     *
+     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
+     * @return this instance
+     */
+    public ContextBuilder requestMessageChannel() {
+        requestMonoList.add(event.getInteraction().getChannel()
+            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
+            .map(messageChannel -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link GuildChannel} the command was called in is required for command execution.
+     * This method will collect the:
+     * messageChannel
+     * guildChannel
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requireTopLevelGuildChannel() {
+        requiredMonoList.add(event.getInteraction().getChannel()
+            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
+            .ofType(TopLevelGuildChannel.class)
+            .doOnNext(guildChannel -> this.topLevelGuildChannel = guildChannel)
+            .map(guildChannel -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link GuildChannel} the command was called in is optionally used for command execution.
+     * This method will collect the:
+     * messageChannel
+     * guildChannel
+     *
+     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
+     * @return this instance
+     */
+    public ContextBuilder requestTopLevelGuildChannel() {
+        requestMonoList.add(event.getInteraction().getChannel()
+            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
+            .ofType(TopLevelGuildChannel.class)
+            .doOnNext(guildChannel -> this.topLevelGuildChannel = guildChannel)
+            .map(guildChannel -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link Member} the command was called by is required for command execution.
+     * This method will collect the:
+     * callingUserAsMember
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requireMember() {
+        requiredMonoList.add(Mono.justOrEmpty(event.getInteraction().getMember())
+            .doOnNext(member -> this.member = member)
+            .map(member -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link Member} the command was called by is optionally used for command execution.
+     * This method will collect the:
+     * callingUserAsMember
+     *
+     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
+     * @return this instance
+     */
+    public ContextBuilder requestMember() {
+        requestMonoList.add(Mono.justOrEmpty(event.getInteraction().getMember())
+            .doOnNext(member -> this.member = member)
+            .map(member -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link User} for the bot is required for command execution.
+     * This method will collect the:
+     * botUser
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requireBotUser() {
+        requiredMonoList.add(event.getClient().getSelf()
+            .doOnNext(user -> this.botUser = user)
+            .map(user -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link User} for the bot is optionally used for command execution.
+     * This method will collect the:
+     * botUser
+     *
+     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
+     * @return this instance
+     */
+    public ContextBuilder requestBotUser() {
+        requestMonoList.add(event.getClient().getSelf()
+            .doOnNext(user -> this.botUser = user)
+            .map(user -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link Member} for the bot is required for command execution.
+     * This method will collect the:
+     * botMember
+     * guild
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requireBotMember() {
+        requiredMonoList.add(event.getInteraction().getGuild()
+            .doOnNext(guild -> this.guild = guild)
+            .flatMap(guild -> guild.getMemberById(event.getClient().getSelfId()))
+            .doOnNext(member -> this.botMember = member)
+            .map(member -> 1));
+        return this;
+    }
+
+    /**
+     * Mark that the {@link Member} for the bot is required for command execution.
+     * This method will collect the:
+     * botMember
+     * guild
+     *
+     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
+     * @return this instance
+     */
+    public ContextBuilder requestBotMember() {
+        requestMonoList.add(event.getInteraction().getGuild()
+            .doOnNext(guild -> this.guild = guild)
+            .flatMap(guild -> guild.getMemberById(event.getClient().getSelfId()))
+            .doOnNext(member -> this.botMember = member)
+            .map(member -> 1));
+        return this;
+    }
 }
