+++ src/main/java/net/exploitables/slashlib/context/UserContextBuilder.java
@@ -1,13 +1,8 @@
 package net.exploitables.slashlib.context;
 
 import discord4j.core.event.domain.interaction.UserInteractionEvent;
-import discord4j.core.object.entity.Guild;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.User;
-import discord4j.core.object.entity.channel.GuildChannel;
-import discord4j.core.object.entity.channel.MessageChannel;
-import discord4j.core.object.entity.channel.TopLevelGuildChannel;
-import reactor.core.publisher.Mono;
 import reactor.util.annotation.NonNull;
 import reactor.util.annotation.Nullable;
 
@@ -20,8 +15,8 @@ public class UserContextBuilder extends ContextBuilder {
     @Nullable User targetUser;
     @Nullable Member targetMember;
 
-    public UserContextBuilder(UserInteractionEvent event) {
-        super(event.getInteraction().getUser());
+    public UserContextBuilder(@NonNull UserInteractionEvent event) {
+        super(event);
 
         this.event          = event;
 
@@ -32,140 +27,6 @@ public class UserContextBuilder extends ContextBuilder {
     @Override
     public UserContext build() {
         return new UserContext(this);
-    }
-
-    /**
-     * Mark that the {@link Guild} the command was called in is required for command execution.
-     * This method will collect the:
-     * guild
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requireGuild() {
-        requiredMonoList.add(event.getInteraction().getGuild()
-            .doOnNext(guild -> this.guild = guild)
-            .map(guild -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link Guild} the command was called in is optionally used for command execution.
-     * This method will collect the:
-     * guild
-     *
-     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requestGuild() {
-        requestMonoList.add(event.getInteraction().getGuild()
-            .doOnNext(guild -> this.guild = guild)
-            .map(guild -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link MessageChannel} the command was called in is required for command execution.
-     * This method will collect the:
-     * messageChannel
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requireMessageChannel() {
-        requiredMonoList.add(event.getInteraction().getChannel()
-            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
-            .map(messageChannel -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link MessageChannel} the command was called in is optionally used for command execution.
-     * This method will collect the:
-     * messageChannel
-     *
-     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requestMessageChannel() {
-        requestMonoList.add(event.getInteraction().getChannel()
-            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
-            .map(messageChannel -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link GuildChannel} the command was called in is required for command execution.
-     * This method will collect the:
-     * messageChannel
-     * guildChannel
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requireTopLevelGuildChannel() {
-        requiredMonoList.add(event.getInteraction().getChannel()
-            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
-            .ofType(TopLevelGuildChannel.class)
-            .doOnNext(guildChannel -> this.topLevelGuildChannel = guildChannel)
-            .map(guildChannel -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link GuildChannel} the command was called in is optionally used for command execution.
-     * This method will collect the:
-     * messageChannel
-     * guildChannel
-     *
-     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requestTopLevelGuildChannel() {
-        requestMonoList.add(event.getInteraction().getChannel()
-            .doOnNext(messageChannel -> this.messageChannel = messageChannel)
-            .ofType(TopLevelGuildChannel.class)
-            .doOnNext(guildChannel -> this.topLevelGuildChannel = guildChannel)
-            .map(guildChannel -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link Member} the command was called by is required for command execution.
-     * This method will collect the:
-     * callingUserAsMember
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requireMember() {
-        requiredMonoList.add(Mono.justOrEmpty(event.getInteraction().getMember())
-            .doOnNext(member -> this.member = member)
-            .map(member -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link Member} the command was called by is optionally used for command execution.
-     * This method will collect the:
-     * callingUserAsMember
-     *
-     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
-     * @return this instance
-     */
-    @Override
-    public UserContextBuilder requestMember() {
-        requestMonoList.add(Mono.justOrEmpty(event.getInteraction().getMember())
-            .doOnNext(member -> this.member = member)
-            .map(member -> 1));
-        return this;
     }
 
     /**
@@ -236,72 +97,6 @@ public class UserContextBuilder extends ContextBuilder {
                 .doOnNext(guild -> this.guild = guild)
                 .flatMap(guild -> guild.getMemberById(user.getId()))
                 .doOnNext(member -> this.targetMember = member))
-            .map(member -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link User} for the bot is required for command execution.
-     * This method will collect the:
-     * botUser
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    public ContextBuilder requireBotUser() {
-        requiredMonoList.add(event.getClient().getSelf()
-            .doOnNext(user -> this.botUser = user)
-            .map(user -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link User} for the bot is optionally used for command execution.
-     * This method will collect the:
-     * botUser
-     *
-     * If the data cannot be collected then {@link Context#doesAllRequestedDataExist()} will return false.
-     * @return this instance
-     */
-    public ContextBuilder requestBotUser() {
-        requestMonoList.add(event.getClient().getSelf()
-            .doOnNext(user -> this.botUser = user)
-            .map(user -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link Member} for the bot is required for command execution.
-     * This method will collect the:
-     * botMember
-     * guild
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    public ContextBuilder requireBotMember() {
-        requiredMonoList.add(event.getInteraction().getGuild()
-            .doOnNext(guild -> this.guild = guild)
-            .flatMap(guild -> guild.getMemberById(event.getClient().getSelfId()))
-            .doOnNext(member -> this.botMember = member)
-            .map(member -> 1));
-        return this;
-    }
-
-    /**
-     * Mark that the {@link Member} for the bot is required for command execution.
-     * This method will collect the:
-     * botMember
-     * guild
-     *
-     * If the data cannot be collected then a {@link DataMissingException} will be thrown.
-     * @return this instance
-     */
-    public ContextBuilder requestBotMember() {
-        requestMonoList.add(event.getInteraction().getGuild()
-            .doOnNext(guild -> this.guild = guild)
-            .flatMap(guild -> guild.getMemberById(event.getClient().getSelfId()))
-            .doOnNext(member -> this.botMember = member)
             .map(member -> 1));
         return this;
     }
