+++ src/test/java/net/exploitables/slashlib/example/basic/interactions/chat/info/InfoGuild.java
@@ -88,6 +88,6 @@ class InfoGuild extends SubCommand {
      */
     @Override
     public ChatContextBuilder setRequestData(ChatContextBuilder contextBuilder) {
-        return contextBuilder.requireGuild();
+        return (ChatContextBuilder) contextBuilder.requireGuild();
     }
 }
