+++ src/test/java/net/exploitables/slashlib/example/basic/interactions/chat/info/entity/InfoChannel.java
@@ -64,6 +64,6 @@ class InfoChannel extends SubCommand {
      */
     @Override
     public ChatContextBuilder setRequestData(ChatContextBuilder contextBuilder) {
-        return contextBuilder.requireTopLevelGuildChannel();
+        return (ChatContextBuilder) contextBuilder.requireTopLevelGuildChannel();
     }
 }
