+++ src/test/java/net/exploitables/slashlib/example/guild/interactions/chat/SyncGuildCommands.java
@@ -31,6 +31,6 @@ public class SyncGuildCommands extends TopCommand {
     }
 
     public ChatContextBuilder setRequestData(ChatContextBuilder builder) {
-        return builder.requireGuild();
+        return (ChatContextBuilder) builder.requireGuild();
     }
 }
