+++ src/test/java/net/exploitables/slashlib/example/guild/interactions/chat/TogglePing.java
@@ -57,6 +57,6 @@ public class TogglePing extends TopCommand {
 
     @Override
     public ChatContextBuilder setRequestData(ChatContextBuilder builder) {
-        return builder.requireGuild();
+        return (ChatContextBuilder) builder.requireGuild();
     }
 }
